refactor(phase-base): extract helpers to Phases/Base
Phase 12b sub-step 3/4: hoist 10 module-private helpers + 4 internal
types out of the 633-LoC Pipeline/Types/BasePhase.ts hub into five
new files under Pipeline/Phases/Base/. The abstract class body stays
in place for one more sub-step so every external importer of
`BasePhase` / `IsPrePayloadValid` keeps working unchanged; the class
itself moves in Commit 4.

New files (each <=110 LoC, each function <=30 LoC -- ESLint section
19.3a budget from the prior commit):

  * PhaseTrace.ts           -- TraceTagStr brand, PHASE_STAGE_EVENT
                               discriminator, RESULT_TAG lookup,
                               traceTag<T>() mapper.
  * BalanceContextSlice.ts  -- BALANCE_SLOT_KEYS, BalanceContextSlice
                               type alias, balanceContextSlice()
                               picker shared by the two builders.
  * BootstrapContextBuilder.ts -- buildBootstrapContext() for INIT and
                               TERMINATE branches (browser, no
                               mediator).
  * ActionContextBuilder.ts -- extractExecutor() + buildActionContext()
                               for the sealed-mediator branch.
  * HandoffHelpers.ts       -- DidEmitHandoff brand, four per-phase
                               resolvers, HANDOFF_MAP OCP registry,
                               PHASE_KEY_MAP normaliser, and the
                               logHandoffSummary() dispatcher.

Pipeline/Types/BasePhase.ts shrinks from 633 -> 399 LoC. It now
imports buildActionContext, logHandoffSummary, PHASE_STAGE_EVENT,
RESULT_TAG, and traceTag from the new sibling modules; the public
`IsPrePayloadValid` brand and the abstract `BasePhase` class stay in
place. Removed imports of extractActionMediator, none, some,
IBootstrapContext and PipelineLogEvent (their consumers all moved).

Validation: tsc --noEmit clean; targeted jest run
(BasePhase|Pipeline/Types|Phases/Init/InitBrowserSetup) -- 329/329
tests pass across 20 suites; tsup rebuild produced
lib/index.d.ts SHA-256 byte-identical to HEAD
(221E1611...AB055, 14823 bytes) -- public surface preserved.
