+++ .github/docs-coverage-allowlist.txt
@@ -465,3 +465,30 @@ SCRAPE_UI_WK_TIMEOUT_MS
 SECOND
 STRONG_AUTH_COOKIE_FLOOR
 TERMINATE_CLEANUP_BUDGET_MS
+
+# --- Phase 12b: BasePhase helper extraction (PR TBD) ---
+# These 12 symbols were previously module-private helpers + brand types
+# defined inside src/Scrapers/Pipeline/Types/BasePhase.ts. Phase 12b
+# sub-step 3/4 hoisted them into five sibling files under
+# src/Scrapers/Pipeline/Phases/Base/ (PhaseTrace.ts,
+# BalanceContextSlice.ts, BootstrapContextBuilder.ts,
+# ActionContextBuilder.ts, HandoffHelpers.ts). They are NOT new public
+# API — they remain Pipeline-internal collaborators of the abstract
+# BasePhase Template Method; lib/index.d.ts SHA-256 stays
+# byte-identical after the move. The docs-coverage scanner sees them
+# as "newly introduced" because it can't follow same-PR cross-file
+# moves. These names were never user-facing; the public surface is the
+# `BasePhase` class and the `IsPrePayloadValid` brand re-exported from
+# Types/BasePhase.ts.
+TraceTagStr
+PHASE_STAGE_EVENT
+RESULT_TAG
+traceTag
+BALANCE_SLOT_KEYS
+BalanceContextSlice
+balanceContextSlice
+buildBootstrapContext
+extractExecutor
+buildActionContext
+DidEmitHandoff
+logHandoffSummary
